+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - Added a new function `tabset()` to represent a list with a tabset. The representation is similar to `str()`, but uses a visual form.
 
+- Factored out the function `taml_load()` and exported it (TAML is a tiny subset of YAML).
+
 - The `print` argument of `record()` can accept non-function values now, in which case `print()` (or `show()` for S4 objects) will be used as the print function.
 
 - The `record()` results can also be formatted to Markdown via `format(record(), to = 'markdown')`.